+++ plume/src/main/kotlin/io/github/plume/oss/drivers/NeptuneDriver.kt
@@ -127,7 +127,6 @@ class NeptuneDriver internal constructor() : GremlinDriver() {
 
     private fun serializeIds() {
         val filePath = "$idStorageLocation$idFileName"
-        println("Serializing IDS $idMapper to $filePath")
         File(filePath).let { f ->
             if (!f.exists())
                 f.createNewFile()
@@ -147,7 +146,10 @@ class NeptuneDriver internal constructor() : GremlinDriver() {
             FileReader(filePath).useLines { ls ->
                 ls.forEach { l ->
                     val split = l.trim().split(":")
-                    if (split.size == 2) idMapper[split[0].toLong()] = split[1]
+                    if (split.size == 2) {
+                        println("Inserting ${split[0].toLong()} -> ${split[1]}")
+                        idMapper[split[0].toLong()] = split[1]
+                    }
                 }
             }
         }